pyproject.toml
[project]
name = "mommap-ai"
version = "0.1.0"
description = "AI microservice for Health Chatbot"
authors = [
    {name = "Theedon",email = "faniyantoyin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "openai (>=1.84.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "redis (>=6.2.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-core (>=0.3.63,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.4.0"
httpx = "^0.28.1"

